Centralize VAD Plugin Testing
After completing the VAD Plugin Testing classes for both plugins,
realized that almost all the code was duplicated between the
two plugins.

Combined almost all the code into the Test_VADPlugin class
located in naomi/testutils.py, which should make it easier to maintain
and simplify the development of new plugins.

I had also added some code which caused the test routine
to output a map of where audio was and was not detected
if the test is run at the info or debug logging levels (the
default logging level for unittests is warn). This had to be
enabled by uncommenting a line on each test if I wanted to
compare the results. With this change, uncommenting or
commenting one line in testutils.py affects the behavior of
both tests.
